SY-4316: Use shared symbolColorVar helper in edges, hoist to schematic level
