+++ pluto/src/schematic/edge/data/spec.tsx
@@ -7,7 +7,7 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
-import { color, type direction, type xy } from "@synnaxlabs/x";
+import { type direction, type xy } from "@synnaxlabs/x";
 import { type ReactElement } from "react";
 
 import { CSS } from "@/css";
@@ -15,6 +15,7 @@ import { Base } from "@/schematic/edge/common/base";
 import { Path } from "@/schematic/edge/common/path";
 import { Segmented } from "@/schematic/edge/common/segmented";
 import { NAME, VARIANT } from "@/schematic/edge/data/config";
+import { symbolColorVar } from "@/schematic/symbolColor";
 
 const SYMBOL_INTERVAL = 40;
 
@@ -39,11 +40,7 @@ export const spec = Segmented.createSpec(
   ({ points, color: colorVal }) => (
     <g
       className={CSS.B("symbol-colored")}
-      style={{
-        [CSS.var("symbol-color")]: color.isZero(colorVal)
-          ? undefined
-          : `${color.rgbString(colorVal)}, ${color.aValue(colorVal)}`,
-      }}
+      style={{ [CSS.var("symbol-color")]: symbolColorVar(colorVal) }}
     >
       <Base.Base path={Path.rounded(points)} color={colorVal} />
       {Path.computeSymbolPositions(points, SYMBOL_INTERVAL).map(
