+++ pluto/src/schematic/node/general/stateIndicator/Primitive.tsx
@@ -16,6 +16,7 @@ import { CSS } from "@/css";
 import { Handle } from "@/schematic/node/common/handle";
 import { Primitive } from "@/schematic/node/common/primitive";
 import { type Config } from "@/schematic/node/general/stateIndicator/config";
+import { symbolColorVar } from "@/schematic/symbolColor";
 import { Text } from "@/text";
 import { Theming } from "@/theming";
 
@@ -45,7 +46,7 @@ export const StateIndicator = ({
   const label = matched != null ? matched.name || `Option ${matched.value}` : "Unknown";
   const style = useMemo<CSSProperties>(
     () => ({
-      [CSS.var("symbol-color")]: Primitive.symbolColorVar(colorVal),
+      [CSS.var("symbol-color")]: symbolColorVar(colorVal),
       backgroundColor,
       minWidth: inlineSize,
     }),
