+++ pluto/src/schematic/symbolColor.ts
